+++ FrontAngular/src/app/app.component.spec.ts
@@ -1,35 +1,35 @@
-import { TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-import { AppComponent } from './app.component';
+// import { TestBed } from '@angular/core/testing';
+// import { RouterTestingModule } from '@angular/router/testing';
+// import { AppComponent } from './app.component';
 
-describe('AppComponent', () => {
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule
-      ],
-      declarations: [
-        AppComponent
-      ],
-    }).compileComponents();
-  });
+// describe('AppComponent', () => {
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       imports: [
+//         RouterTestingModule
+//       ],
+//       declarations: [
+//         AppComponent
+//       ],
+//     }).compileComponents();
+//   });
 
-  it('should create the app', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    expect(app).toBeTruthy();
-  });
+//   it('should create the app', () => {
+//     const fixture = TestBed.createComponent(AppComponent);
+//     const app = fixture.componentInstance;
+//     expect(app).toBeTruthy();
+//   });
 
-  it(`should have as title 'itemsProcess'`, () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    expect(app.title).toEqual('itemsProcess');
-  });
+//   it(`should have as title 'itemsProcess'`, () => {
+//     const fixture = TestBed.createComponent(AppComponent);
+//     const app = fixture.componentInstance;
+//     expect(app.title).toEqual('itemsProcess');
+//   });
 
-  it('should render title', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('.content span')?.textContent).toContain('itemsProcess app is running!');
-  });
-});
+//   it('should render title', () => {
+//     const fixture = TestBed.createComponent(AppComponent);
+//     fixture.detectChanges();
+//     const compiled = fixture.nativeElement as HTMLElement;
+//     expect(compiled.querySelector('.content span')?.textContent).toContain('itemsProcess app is running!');
+//   });
+// });
