+++ FrontAngular/src/app/shared/index.ts
@@ -1,1 +1,1 @@
-export * from './models';
+export * from './models/';
