+++ FrontAngular/src/app/shared/models/cliente.model.ts
@@ -10,15 +10,4 @@
 }*/
 
 import { Cliente } from "./cliente";
-import { ItemDoPedido } from "./itemdopedido.model";
 
-
-export class Pedido{
-    
-    constructor(public date: Date | String,public cliente: Cliente,public itens: ItemDoPedido[]){
-        this.date = date
-        this.cliente = cliente;   
-        this.itens = itens;
-    }
-
-}
