+++ FrontAngular/src/app/shared/models/index.ts
@@ -1,1 +1,1 @@
-export * from './cliente.model';
+export * from './';
