+++ FrontAngular/src/app/shared/models/itemdopedido.model.spec.ts
@@ -1,7 +1,7 @@
-import { ItemDoPedido } from './itemdopedido.model';
+// import { ItemDoPedido } from './itemdopedido.model';
 
-describe('Itemdopedido', () => {
-  it('should create an instance', () => {
-    expect(new ItemDoPedido()).toBeTruthy();
-  });
-});
+// describe('Itemdopedido', () => {
+//   it('should create an instance', () => {
+//     expect(new ItemDoPedido()).toBeTruthy();
+//   });
+// });
