+++ FrontAngular/src/app/shared/models/itemdopedido.model.ts
@@ -1,4 +1,3 @@
-import { Pedido } from ".";
 import { Cliente } from "./cliente";
 
 export interface ItemDoPedido{
