+++ FrontAngular/src/app/views/cliente/editar-cliente/editar-cliente.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { EditarClienteComponent } from './editar-cliente.component';
+// import { EditarClienteComponent } from './editar-cliente.component';
 
-describe('EditarPessoaComponent', () => {
-  let component: EditarClienteComponent;
-  let fixture: ComponentFixture<EditarClienteComponent>;
+// describe('EditarPessoaComponent', () => {
+//   let component: EditarClienteComponent;
+//   let fixture: ComponentFixture<EditarClienteComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ EditarClienteComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ EditarClienteComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EditarClienteComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(EditarClienteComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
