+++ FrontAngular/src/app/views/cliente/inserir-cliente/inserir-cliente.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { InserirClienteComponent } from './inserir-cliente.component';
+// import { InserirClienteComponent } from './inserir-cliente.component';
 
-describe('InserirPessoaComponent', () => {
-  let component: InserirClienteComponent;
-  let fixture: ComponentFixture<InserirClienteComponent>;
+// describe('InserirPessoaComponent', () => {
+//   let component: InserirClienteComponent;
+//   let fixture: ComponentFixture<InserirClienteComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ InserirClienteComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ InserirClienteComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(InserirClienteComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(InserirClienteComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
