+++ FrontAngular/src/app/views/cliente/listar-cliente/listar-cliente.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ListarClienteComponent } from './listar-cliente.component';
+// import { ListarClienteComponent } from './listar-cliente.component';
 
-describe('ListarPessoaComponent', () => {
-  let component: ListarClienteComponent;
-  let fixture: ComponentFixture<ListarClienteComponent>;
+// describe('ListarPessoaComponent', () => {
+//   let component: ListarClienteComponent;
+//   let fixture: ComponentFixture<ListarClienteComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ ListarClienteComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ ListarClienteComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ListarClienteComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(ListarClienteComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
