+++ FrontAngular/src/app/views/cliente/remover-cliente/remover-cliente.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { RemoverClienteComponent } from './remover-cliente.component';
+// import { RemoverClienteComponent } from './remover-cliente.component';
 
-describe('RemoverClienteComponent', () => {
-  let component: RemoverClienteComponent;
-  let fixture: ComponentFixture<RemoverClienteComponent>;
+// describe('RemoverClienteComponent', () => {
+//   let component: RemoverClienteComponent;
+//   let fixture: ComponentFixture<RemoverClienteComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ RemoverClienteComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ RemoverClienteComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(RemoverClienteComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(RemoverClienteComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
