+++ FrontAngular/src/app/views/cliente/services/cliente.service.spec.ts
@@ -1,8 +1,7 @@
 import { TestBed } from '@angular/core/testing';
+import { ClienteService } from './cliente.service'
 
-import { ClienteService } from './cliente.service';
-
-describe('PessoaService', () => {
+describe('ClienteService', () => {
   let service: ClienteService;
 
   beforeEach(() => {
