+++ FrontAngular/src/app/views/cliente/services/cliente.service.ts
@@ -1,6 +1,4 @@
 import { Injectable } from '@angular/core';
-
-
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 
 imports: [
