+++ FrontAngular/src/app/views/pedido/inserir-pedido/inserir-pedido.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, NgForm } from '@angular/forms';
 import { MatDialogRef } from '@angular/material/dialog';
 import { Router } from '@angular/router';
-import { Pedido } from 'src/app/shared';
+import { Pedido } from 'src/app/shared/models/cliente';
 import { Cliente } from 'src/app/shared/models/cliente';
 import { ItemDoPedido } from 'src/app/shared/models/itemdopedido.model';
 import { Produto } from 'src/app/shared/models/produto.model';
