+++ FrontAngular/src/app/views/pedido/listar-pedido/listar-pedido.component.ts
@@ -3,7 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute } from '@angular/router';
 import { Cliente } from 'src/app/shared/models/cliente';
-import { Pedido } from 'src/app/shared/models/cliente.model';
+import { Pedido } from 'src/app/shared/models/cliente';
 import { ItemDoPedido } from 'src/app/shared/models/itemdopedido.model';
 import { Produto } from 'src/app/shared/models/produto.model';
 import { InserirPedidoComponent } from '../inserir-pedido/inserir-pedido.component';
