+++ FrontAngular/src/app/views/pedido/services/pedido.service.spec.ts
@@ -1,16 +1,16 @@
-import { TestBed } from '@angular/core/testing';
+//import { TestBed } from '@angular/core/testing';
 
-import { ServicesService } from './pedido.service';
+//import { ServicesService } from './pedido.service';
 
-describe('ServicesService', () => {
-  let service: ServicesService;
+//describe('ServicesService', () => {
+//  let service: ServicesService;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ServicesService);
-  });
+//  beforeEach(() => {
+//    TestBed.configureTestingModule({});
+//    service = TestBed.inject(ServicesService);
+//  });
 
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
+//  it('should be created', () => {
+//    expect(service).toBeTruthy();
+//  });
+//});
