+++ FrontAngular/src/app/views/produto/editar-produto/editar-produto.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { EditarProdutoComponent } from './editar-produto.component';
+// import { EditarProdutoComponent } from './editar-produto.component';
 
-describe('EditarProdutoComponent', () => {
-  let component: EditarProdutoComponent;
-  let fixture: ComponentFixture<EditarProdutoComponent>;
+// describe('EditarProdutoComponent', () => {
+//   let component: EditarProdutoComponent;
+//   let fixture: ComponentFixture<EditarProdutoComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ EditarProdutoComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ EditarProdutoComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EditarProdutoComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(EditarProdutoComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
