+++ FrontAngular/src/app/views/produto/inserir-produto/inserir-produto.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { InserirProdutoComponent } from './inserir-produto.component';
+// import { InserirProdutoComponent } from './inserir-produto.component';
 
-describe('InserirProdutoComponent', () => {
-  let component: InserirProdutoComponent;
-  let fixture: ComponentFixture<InserirProdutoComponent>;
+// describe('InserirProdutoComponent', () => {
+//   let component: InserirProdutoComponent;
+//   let fixture: ComponentFixture<InserirProdutoComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ InserirProdutoComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ InserirProdutoComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(InserirProdutoComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(InserirProdutoComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
