+++ FrontAngular/src/app/views/produto/listar-produto/listar-produto.component.spec.ts
@@ -1,25 +1,25 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ListarProdutoComponent } from './listar-produto.component';
+// import { ListarProdutoComponent } from './listar-produto.component';
 
-describe('ListarProdutoComponent', () => {
-  let component: ListarProdutoComponent;
-  let fixture: ComponentFixture<ListarProdutoComponent>;
+// describe('ListarProdutoComponent', () => {
+//   let component: ListarProdutoComponent;
+//   let fixture: ComponentFixture<ListarProdutoComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ ListarProdutoComponent ]
-    })
-    .compileComponents();
-  });
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [ ListarProdutoComponent ]
+//     })
+//     .compileComponents();
+//   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ListarProdutoComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(ListarProdutoComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
